refactort(list): 멘토님 피드백 반영
-셀렉터 외부 클릭 시 사라지는 거 훅으로 빼기
-셀렉터가 보내주는 값  dataset쓰지말고 상태로 사용하기
-이벤트 리스너 li에 직접 달아주기
