+++ src/components/postId/AnswerCluster.jsx
@@ -1,11 +1,11 @@
 import { useEffect, useRef } from 'react';
 
+import Empty from '@assets/icons/Empty.svg?react';
+import Message from '@assets/icons/Messages.svg?react';
 import useIntersectionObserver from '@hooks/useIntersectionObserver';
 import styled from 'styled-components';
 
 import AnswerItem from './AnswerItem';
-import Empty from '../../assets/Empty.svg?react';
-import Message from '../../assets/Messages.svg?react';
 
 function AnswerCluster({ subjectInfo, result, questions, callback }) {
   const loadingRef = useRef(null);
